requirements.txt
beautifulsoup4==4.9.0
blis==0.4.1
catalogue==1.0.0
certifi==2020.4.5.1
chardet==3.0.4
cymem==2.0.3
es-core-news-sm==2.2.5
idna==2.9
importlib-metadata==1.6.0
lxml==4.5.0
murmurhash==1.0.2
numpy==1.18.3
Pillow==7.1.2
plac==1.1.3
preshed==3.0.2
python-dotenv==0.13.0
python-pptx==0.6.18
requests==2.23.0
soupsieve==2.0
spacy==2.2.4
srsly==1.0.2
thinc==7.4.0
tqdm==4.45.0
urllib3==1.25.9
wasabi==0.6.0
XlsxWriter==1.2.8
zipp==3.1.0
